feat: use fuzzy search for command palette
When typing in workflow names using the command palette, I often
find myself wanting Sublime Text like fuzzy matching so I don't
have to remember the full, exact name of the workflow.

This PR adds the fuzzaldrin-plus package that is the successor of
the package used in the Atom editor to do fuzzy filtering.

* https://github.com/jeancroy/fuzz-aldrin-plus
* https://github.com/atom/fuzzaldrin